fix モデルを別ファイルに切り出し
